FEAT: 기록 리프레시

+++ src/main/java/com/workout/heavylift/domain/WorkoutLogExercise.java
@@ -25,11 +25,11 @@ public class WorkoutLogExercise {
 
     private int sets;
     private int reps;
-    private int weight;
-    private int totalWeight;
+    private double weight;
+    private double totalWeight;
 
     @Builder
-    public WorkoutLogExercise(WorkoutLog workoutLog, Exercise exercise, int sets, int reps, int weight,int totalWeight) {
+    public WorkoutLogExercise(WorkoutLog workoutLog, Exercise exercise, int sets, int reps, double weight,double totalWeight) {
         this.workoutLog = workoutLog;
         this.exercise = exercise;
         this.sets = sets;
@@ -38,7 +38,7 @@ public class WorkoutLogExercise {
         this.totalWeight = totalWeight;
     }
 
-    public void updateLogExercise(int sets, int reps, int weight) {
+    public void updateLogExercise(int sets, int reps, double weight) {
         this.sets = sets;
         this.reps = reps;
         this.weight = weight;
